# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/member/quotes/{quoteId}/orders": {
      "post": {
        "description": "Creates the Order that accepts a Quote. No money moves at creation; only Confirm has a financial effect.",
        "operationId": "MemberOrderController_create",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "description": "Scoped Idempotency-Key. Replaying it returns the same Bet Order; the same key with a different Quote conflicts.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetOrderBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a Bet Order from an authorised Quote",
        "tags": [
          "Member Betting Orders"
        ]
      }
    },
    "/api/v1/member/orders/{id}": {
      "get": {
        "operationId": "MemberOrderController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetOrderBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read a Member Bet Order by id",
        "tags": [
          "Member Betting Orders"
        ]
      }
    },
    "/api/v1/member/orders/{id}/confirm": {
      "post": {
        "description": "Revalidates the Quote and the authoritative Draw cutoff, then requires the durable Wallet & Ledger stake reserve/commit before the Order becomes CONFIRMED. Denials resolve to REJECTED with no money moved.",
        "operationId": "MemberOrderController_confirm",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "description": "Scoped Idempotency-Key for the Confirm command.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BetOrderCommandBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetOrderBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm a Bet Order (explicit command)",
        "tags": [
          "Member Betting Orders"
        ]
      }
    },
    "/api/v1/member/orders/{id}/cancel": {
      "post": {
        "description": "A Member may cancel before the Draw cutoff. The Order becomes CANCELLED only after the refund posting is durable.",
        "operationId": "MemberOrderController_cancel",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "description": "Scoped Idempotency-Key for the Cancel command.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BetOrderCommandBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetOrderBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a confirmed Bet Order (explicit command)",
        "tags": [
          "Member Betting Orders"
        ]
      }
    },
    "/api/v1/member/orders/{id}/receipt": {
      "get": {
        "operationId": "MemberOrderController_receipt",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetReceiptBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the immutable Bet Receipt of a confirmed Order",
        "tags": [
          "Member Betting Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lottify v1 API",
    "description": "Authoritative REST/OpenAPI contract for Lottify v1",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BetOrderLineBody": {
        "type": "object",
        "properties": {
          "betTypeId": {
            "type": "string"
          },
          "betTypeCode": {
            "type": "string"
          },
          "betTypeVersionId": {
            "type": "string"
          },
          "canonicalNumber": {
            "type": "string",
            "example": "42"
          },
          "stakeMinor": {
            "type": "string",
            "description": "Stake in integer minor units"
          },
          "resolvedPayout": {
            "type": "object",
            "description": "Server-resolved payout accepted for the line (opaque configuration value)"
          },
          "payoutSource": {
            "type": "string",
            "enum": [
              "DRAW_OVERRIDE",
              "DRAW_SNAPSHOT"
            ]
          },
          "restrictions": {
            "description": "Accepted restriction codes",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "betTypeId",
          "betTypeCode",
          "betTypeVersionId",
          "canonicalNumber",
          "stakeMinor",
          "resolvedPayout",
          "payoutSource",
          "restrictions"
        ]
      },
      "BetOrderBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "memberId": {
            "type": "string"
          },
          "quoteId": {
            "type": "string"
          },
          "drawId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "productVersionId": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "enum": [
              "THB"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "DRAFT",
              "QUOTED",
              "CONFIRMING",
              "CONFIRMED",
              "CANCELLING",
              "CANCELLED",
              "EXPIRED",
              "REJECTED",
              "SETTLED"
            ]
          },
          "version": {
            "type": "number",
            "description": "Optimistic-concurrency version"
          },
          "allowedActions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CONFIRM",
                "CANCEL"
              ]
            }
          },
          "totalStakeMinor": {
            "type": "string"
          },
          "cutoffAt": {
            "type": "string",
            "format": "date-time"
          },
          "quoteExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "reservationId": {
            "type": "string",
            "nullable": true
          },
          "stakeTransactionId": {
            "type": "string",
            "nullable": true
          },
          "refundTransactionId": {
            "type": "string",
            "nullable": true
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          },
          "confirmedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "rejectedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "receiptId": {
            "type": "string",
            "nullable": true,
            "description": "Issued Receipt id"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BetOrderLineBody"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "memberId",
          "quoteId",
          "drawId",
          "productId",
          "productVersionId",
          "currency",
          "state",
          "version",
          "allowedActions",
          "totalStakeMinor",
          "cutoffAt",
          "quoteExpiresAt",
          "reservationId",
          "stakeTransactionId",
          "refundTransactionId",
          "rejectionReason",
          "cancellationReason",
          "confirmedAt",
          "cancelledAt",
          "rejectedAt",
          "receiptId",
          "lines",
          "createdAt",
          "updatedAt"
        ]
      },
      "BetOrderCommandBody": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "description": "Expected Bet Order version"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "version"
        ]
      },
      "BetReceiptLineBody": {
        "type": "object",
        "properties": {
          "betTypeCode": {
            "type": "string"
          },
          "betTypeVersionId": {
            "type": "string"
          },
          "canonicalNumber": {
            "type": "string"
          },
          "stakeMinor": {
            "type": "string"
          },
          "resolvedPayout": {
            "type": "object",
            "description": "Payout accepted at confirmation (opaque configuration value)"
          }
        },
        "required": [
          "betTypeCode",
          "betTypeVersionId",
          "canonicalNumber",
          "stakeMinor",
          "resolvedPayout"
        ]
      },
      "BetReceiptTermsBody": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "productVersionId": {
            "type": "string"
          },
          "drawReference": {
            "type": "string",
            "nullable": true
          },
          "drawCutoffAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "enum": [
              "THB"
            ]
          },
          "totalStakeMinor": {
            "type": "string"
          },
          "acceptedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BetReceiptLineBody"
            }
          },
          "acceptedRestrictions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "productId",
          "productVersionId",
          "drawReference",
          "drawCutoffAt",
          "currency",
          "totalStakeMinor",
          "acceptedAt",
          "lines",
          "acceptedRestrictions"
        ]
      },
      "BetReceiptBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "memberId": {
            "type": "string"
          },
          "orderVersion": {
            "type": "number"
          },
          "contentDigest": {
            "type": "string",
            "description": "SHA-256 over the accepted terms"
          },
          "terms": {
            "$ref": "#/components/schemas/BetReceiptTermsBody"
          },
          "issuedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "orderId",
          "memberId",
          "orderVersion",
          "contentDigest",
          "terms",
          "issuedAt"
        ]
      },
...
    }
  }
}